refactor(rag): improve code structure
