+++ src/pages/graph-rag/settings/index.ts
@@ -1,0 +1,1 @@
+export { GraphRAGSettingsDialog } from "./GraphRAGSettingsDialog";
