+++ src/pages/graph-rag/viewer/NodeStats.tsx
